Move abstract classes for Local tasks up one folder and rename slightly
