+++ classes/suggested-tasks/local-tasks/providers/one-time/class-remove-inactive-plugins.php
@@ -7,6 +7,8 @@
 
 namespace Progress_Planner\Suggested_Tasks\Local_Tasks\Providers\One_Time;
 
+use Progress_Planner\Suggested_Tasks\Local_Tasks\Providers\One_Time;
+
 /**
  * Add tasks to check if WP debug is enabled.
  */
